refactor(broker): split creating server transport out of command api
The transport is used by commandApi, adminApi and backupApi. Since it is a shared resource, it is better not to tie it with command api step.
